pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ticket-triage-env"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0",
    "openai>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ticket_triage_env*", "server*"]
